+++ Changelog.md
@@ -1,5 +1,18 @@
 # Changelng
 
+## 29 April 2012
+
+* Ensured that the timepicker has the correct position when it fades
+  in. [Bug fix]
+
+## 17 April 2012
+
+* Added VERSION
+
+## 16 April 2012
+
+* Added code to allow the timepicker to be added into a popup container
+
 ## 05 April 2012
 
 * Updated README